examples/python-ml/requirements.txt
numpy==1.21.0
pillow==8.3.1
flask==1.0.2
requests==2.20.0
pyyaml==5.3
scikit-learn==0.24.0
pandas==1.2.0
